refactor :  기능구현 부분과 기능수행 부분을 별도의 클래스로 분리
RacingCarGame 클래스에 기능수행 관련 메소드 구현
RacingCar 클래스에 기능구현 관련 메소드 구현
Application 클래스에 RacingCarGame을 실행하는 메소드 구현
