+++ src/main/java/racingcar/Application.java
@@ -3,5 +3,7 @@ package racingcar;
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
+        RacingCarGame racingCar = new RacingCarGame();
+        racingCar.playRacingCar();
     }
 }
